Moved ProjectRules to its own folder
