+++ src/components/project/index.ts
@@ -2,5 +2,5 @@ export * from './dto';
 export * from './project.service';
 export * from './project-member';
 export * from './list-filter.query';
-export * from './project.rules';
+export * from './project-rules/project.rules';
 export * from './project.loader';
